subscription revenue calculated on admin dashboard

+++ src/main/java/com/tlt/entities/SubscriptionModel.java
@@ -5,7 +5,6 @@
 package com.tlt.entities;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Collection;
 import javax.persistence.Basic;
 import javax.persistence.Column;
@@ -22,7 +21,7 @@ import javax.validation.constraints.Size;
 
 /**
  *
- * @author Lenovo
+ * @author kunal
  */
 @Entity
 @Table(name = "subscription_model")
@@ -47,9 +46,8 @@ public class SubscriptionModel implements Serializable {
     private String name;
     @Column(name = "duration_in_month")
     private Integer durationInMonth;
-    // @Max(value=?)  @Min(value=?)//if you know range of your decimal fields consider using these annotations to enforce field validation
     @Column(name = "cost")
-    private BigDecimal cost;
+    private Integer cost;
     @Lob
     @Size(max = 65535)
     @Column(name = "details")
@@ -88,11 +86,11 @@ public class SubscriptionModel implements Serializable {
         this.durationInMonth = durationInMonth;
     }
 
-    public BigDecimal getCost() {
+    public Integer getCost() {
         return cost;
     }
 
-    public void setCost(BigDecimal cost) {
+    public void setCost(Integer cost) {
         this.cost = cost;
     }
